pyproject.toml
[project]
name = "clintela"
version = "0.1.0"
description = "AI-powered post-surgical patient recovery support"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "TBD"}
authors = [
    {name = "Clintela Team", email = "team@clintela.com"},
]
keywords = ["healthcare", "ai", "patient-care", "django"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: Django :: 5.1",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Django & Core
    "django>=5.1,<5.2",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",

    # Database
    "psycopg2-binary>=2.9.9",

    # Real-time (WebSockets)
    "channels>=4.0",
    "daphne>=4.0",
    "channels-redis>=4.2",

    # AI/LLM
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.0.50",

    # Communications
    "twilio>=9.0",

    # HTTP/API
    "httpx>=0.27",
    "django-ninja>=1.0",

    # Security
    "argon2-cffi>=23.1",
    "cryptography>=42.0",
    "django-ratelimit>=4.1.0",

    # Utilities
    "pydantic>=2.7",
    "python-dateutil>=2.9",
    "phonenumbers>=8.13",
    "bleach>=6.1",

    # Async
    "asgiref>=3.8",

    # Background Tasks
    "celery[redis]>=5.4",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.2",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.5",
    "factory-boy>=3.3",
    "faker>=25.0",

    # Linting & Formatting
    "ruff>=0.4.0",

    # Type Checking
    "django-stubs>=5.0",
    "mypy>=1.10",

    # Development Tools
    "pre-commit>=3.7",
    "django-debug-toolbar>=4.4",
    "ipython>=8.24",
    "ipdb>=0.13",

    # Security
    "bandit>=1.7",
    "detect-secrets>=1.5",

    # Documentation
    "pytest-md>=0.2.0",

    # E2E Testing
    "pytest-playwright>=0.5.0",
    "playwright>=1.40",
]

prod = [
    # Production Server
    "gunicorn>=22.0",
    "whitenoise>=6.6",

    # Monitoring
    "sentry-sdk>=2.0",
    "structlog>=24.1",

    # Performance
    "django-redis>=5.4",
]

[project.scripts]
clintela = "manage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "config/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]

# UV Configuration
[tool.uv.pip]
generate-hashes = true

# Ruff Configuration
[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4
exclude = [
    "migrations",
    "static",
    "templates",
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "DJ",  # flake8-django
    "S",   # flake8-bandit (security)
    "C90", # mccabe complexity
]
ignore = [
    "S101",  # assert statements in tests are fine
    "S105",  # hardcoded passwords in tests are fine
    "S106",  # hardcoded passwords in tests are fine
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

# Pytest Configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests", "apps"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
    "--color=yes",
    "--ignore=node_modules",
    "--ignore=tests/e2e",
    "--numprocesses=auto",
    "--dist=worksteal",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end browser tests (run with: pytest tests/e2e/ -p no:xdist)",
]

# Coverage Configuration
[tool.coverage.run]
source = ["apps", "config"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "config/settings/*",
    "config/wsgi.py",
    "config/asgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 90
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

# Mypy Configuration
[tool.mypy]
python_version = "3.12"
django_settings_module = "config.settings.development"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "twilio.*",
    "langchain.*",
    "langgraph.*",
]
ignore_missing_imports = true

# Bandit Configuration
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "__pycache__"]
skips = ["B101"]  # assert_used in tests

[tool.bandit.assert_used]
excludes = ["tests/*", "*/tests/*"]
